CI configs: Keep disabled CI configuration files around.
In order to speed up CIs and avoid long queues, we decided to remove CI
files if the package was not configured. While this is a great system
and definitely helps, it increases the barrier (slightly) to adding back
support. Namely, we need to add this file back. This could be
intimidating to a new user or someone just trying to get something
working quickly. For instance, they might worry that portions of these
files are unique to each feedstock.

This proposal suggests that instead of deleting these files entirely, we
just rename them. That way the file remains around and CI remains
disabled. However, if we want to try quickly to re-enable this
functionality we simply move the file back.