CI: Add the conda-team channel.
This channel is where development builds of `conda-build` are being
placed. By adding it to our channels, this will give us a better chance
at trying out new features or bug fixes in `conda-build` that are not
official released. Hopefully by doing this, we can work with Continuum
to iron out any issues before an official release. Also, this should
help us try out things that believe are potential fixes.