CI Tmplts: Do not make `pip` and `setuptools` dependencies of `python`.
This disables adding `pip` and `setuptools` as dependencies of `python`.
By doing this, it should make it easier for us to figure out if
something requires `setuptools` to install as it will error out if
`setuptools` is not listed as a build dependency. Similar, nothing will
be able to try to use `pip` to install stuff. This will make it easier
to figure out what dependencies we are missing on a package. Further,
this will be helpful for packaging `pip` and `setuptools` too.