Change the check or upload script to use the index.
Likely our request for the file list before was overwhelming the API
as it had to collect all this info for 1000s of packages. This should
hopefully be a much easier thing to get as the index file should already
exist (read does not need to be generated by the API).