Build as a non-root user
This requires the new docker image which creates a non-root user

As we don't run as root anymore, the conda-build process won't be
able to write to the mounted volumes unless the uid inside the
container is the same as the host (assuming the user has permission
to write the mounted volumes in the host side).

When using docker-machine, we need to get the uid from the virtual
machine created by docker-machine.
Thanks to @jakirkham for discovering that.

Fix #212