Handle a subtle case during re-rendering involving file removal.
In some cases a file may already be removed from the file system.
However, the corresponding change may not have been added to the index.
We were not able to handle this case previously as GitPython is unable
to remove the file from the index if it is not already on the file
system.

So we had to check that the file was on the file system first before
trying to remove it from the index. The result being a file removed from
the file system and not removed from the index could linger in the
subsequent commit even if it should have been removed. This issue could
crop up if someone tried to re-render the feedstock (resulting in some
files being removed). Then backed out the changes from the index with
`git reset` for some reason.

To solve this issue, this change creates an empty file on the file
system and stages it to the index before trying to remove it. By doing
this, we ensure to always have the file on the file system and in the
index beforehand. Thus removal is always possible from both without
issue.