Prune empty directories.
If a file that we removed was the last file in a directory, remove that
directory and remove any and all parent directories that are also empty.
However, as soon as a non-empty directory is encountered, stop trying to
remove directories. This should help avoid empty directories lingering
that `git` would not otherwise track.