Set `CONDA_PY` on AppVeyor if undefined.
This is done as it is used by the Obvious-CI compiler activation script
for AppVeyor ( `obvci_appveyor_python_build_env.cmd` ). That script
expects to get `TARGET_ARCH` and `CONDA_PY`. We already do a good job of
setting `TARGET_ARCH` in the environment matrix. However, there was one
case where `CONDA_PY` was not set. Namely if `python` was not a build
dependency. Here we fix that to restore the old behavior of using Python
2.7 as the default, but we do this in the matrix to make it a bit
clearer what it is we are doing.