Fix source channel order in CIs.
The order the source channels were listed in `conda-forge.yml` was not
being respected. In the case of Travis CI and AppVeyor, these were
getting added in the wrong order so the priority would have been
backwards (i.e. last channel in `conda-forge.yml` would have come
first). In the CircleCI/docker build, the channels were being sorted,
which should not occur. Thus order was not being respected at all in
that case.

Here we reverse source channel additions in the Travis CI and AppVeyor
configuration files so that will get the right priority. In the docker
build configuration, we drop the unneeded sorting.