Override the definition of `defaults`.
When getting the index, `conda` has a notion of `defaults` (at least in
older versions) that is dependent on the platform it is run on.
Unfortunately, this may result in using things like the `msys2` channel
when re-rendering on Linux even though there are no Linux packages in
this case.

To fix this we add a hack that detects the `defaults` channel if listed
and redefines it based on the platform we are masquerading as. This
hopefully should allow Windows users to re-render natively. Also this
should ensure that we do use `msys2` on Windows so that feedstocks
requiring it will be re-rendered properly.