Disable execute permissions on newly written files
In the case of some file systems, permissions are not set or do not show
up correctly. In particular, it has been demonstrated in a few cases all
files show up with executable permissions. To handle this issue, we
explicitly disable execute permissions on all newly written files and we
make sure this gets added to `git`'s index. Adding this to `git`'s index
is necessary as the file system likely will still show the new files
with execute permission even though we have tried to disable them.