Preserve file's original line endings
Instead of converting all newlines to `\n`, preserve line endings of
files as they are on disk. This should be a nicer solution than simply
forcing newlines to `\n` in all cases when writing file as it can keep
files with `\r\n` as is. Also this should still work nicely on Windows.