Drop Python 3.4 AppVeyor update hack remnant
This remnant was part of a hack to make sure AppVeyor's preincluded
Miniconda for Python 3.4 had a new enough version of `conda`. The
included version could not be activated properly so this was necessary
to make it work. However, we have since dropped support for Python 3.4.
Further the preincluded Python 3.5 Miniconda has a very recent version
of `conda`. As a result, this hack is no longer necessary and can be
dropped.