Fast finish support for conda-forge-build-setup
Change the fast finish support to special case for
`conda-forge-build-setup`. This should allow for the fast finish scripts
to be used directly from `conda-forge-build-setup` instead of trying to
download them from `master` branch. By doing this, it will be easier to
test changes made to them as part of `conda-forge-build-setup`'s CI.