Simplify feedstock repo initialization
As we already stage changes to a feedstock's git repo (should it exist),
it makes sense that we actually create a git repo if we will need one
before we add content to the feedstock. By doing this, we can simplify
and cut redundant code from `conda_smithy.cli` that is already handled
effectively in `conda_smithy.configure_feedstock`. Thus reducing our
chances of code rot and also avoiding contaminating any other repo.