Optionally handle team additions
If specified by a command line argument, add a team for the maintainers
of the feedstock in the org (if it is an org). Also add a team for all
of the maintainers and add the maintainers of the current feedstock to
that team. This is based off of code in the webpage repo and
staged-recipes to handle this same sort of addition. It is being added
for the intent of consolidating the code from both locations. Plus it
should open the door for a supporting web services that handles team
additions asynchronously.