conda_smithy/templates/run_docker_build_matrix.tmpl: don't toggle logging for matrix environment
The old version had some accidentally inverted `set +-x` flags around the
stanzas that set up the environment variables for each build matrix run. It
looks like these were put in to mirror the bit of shell script that prevents
echoing of the secret Binstar token, but it's useful to see the variables
here. So let's just not toggle the shell's `x` flag at all.