Listen to repository events
The repository event notifies when a repo has been deleted (amongst
other things). This is important for the team and feedstocks services as
they will also need to handle this event. The team service will need to
remove the associated team and the feedstocks service will need to
remove it from the feedstocks repo.