Add an issue template for feedstocks
To better format issues raised going forward and help resolve them more
quickly, this includes an issue template that GitHub will use for any
new issues created. The template requests simple information about
demonstrating a problem users encounter and providing info about the
environment that created the problem so it can be reproduced. Also the
template requests info about the `conda` install and system info. As
these questions are asked regularly of new issues reported (often some
time after the report), it would be better both for users and
maintainers to get this information into the report at the time it is
opened. Also this information is just as handy for feature requests as
bug reports because it can inform tests that can be included into the
build to ensure the problem is resolved and stays that way.