Drop the `--no-git-repo` option
The `--no-git-repo` option is unused in `conda-forge`. Also it doesn't
sound like it is being used outside of `conda-forge`. Most of the
content that `conda-smithy` generates ends up being only useful for
GitHub repos with CI configured anyways. In an effort to help simplify
the codebase, this drops the `--no-git-repo` option. Hopefully things
like `feedstock_io` can be simplified in the process.