Fix Windows badge in feedstocks' README
There are some tweaks needed for the repo name for the AppVeyor/Windows
badge. While these were included in the link, it seems they need to be
included in the Shields.io badge as well for the badge to render. Should
fix some issues with the rendering of AppVeyor/Windows badges on
feedstocks that have `_`s or `.`s in their names.