Move the NumFOCUS badge to "About conda-forge"
As there was some concern raised that having the NumFOCUS badge where it
was would confuse users as to what was NumFOCUS sponsored (i.e. that
their library is NumFOCUS sponsored) as opposed to the fact that
conda-forge is NumFOCUS sponsored, this relocates the NumFOCUS badge to
the "About conda-forge" section where this should hopefully make it
clear that conda-forge is NumFOCUS sponsored.