Remove `/opt/ghc` on Linux
The content in `/opt/ghc` seems to take up a fair bit of space on the
Linux images and we don't use it. As it limits the memory available for
jobs that are more memory hungry and doesn't seem to take long to clear,
go ahead and remove `/opt/ghc` at the start of the job.