Use 50/72 rule with git commit messages
Updates conda-smithy's commit messages and recommended commit messages
to follow the recommended 50/72 rule. Namely it keeps the subject line
to 50 characters and the body of the commit to 72 characters per line
(wrapping when exceeding that length) with a blank line between. GitHub
also uses this rule when displaying commit messages by effectively
hiding anything over 50 characters in the subject line and hiding the
body of the commit message.