Use build ID instead of job ID to name Azure artifacts, and truncate.
The job ID is pretty long whereas the build ID is short, and if the
artifact name is too long (~100 characters? I could not find docs) then
it will not be downloadable. This truncates the artifact name to 80
characters (using the shell to mangle the variable, because I don't
think there's a better way to do this with Azure functions) to prevent
the artifact from not being downloadable.