Add option to use different path for feedstock config yaml.
This makes it possible to specify the feedstock configuration in a file
that is named/located at something other than "conda-forge.yml".
Feedstock init and ci-skeleton still use the default name, but if a user
decides to manually move/rename the config file, there is now a
"--feedstock_config" option for appropriate smithy subcommands so that
they can point to the alternate path.

One test has been added for configure_feedstock to ensure that the
configuration can be loaded from an alternative path.