skip_render can skip github_actions_services
Also enhanced it so that you can skip rendering anything
below a directory by specifying the directory name (without
For example, if '.github' is in the skip_render list, all
files inside .github/ will be skipped.

Logging now tells you when rendering was actually skipped
as opposed to just what was read from the config.