Fix syntax error when preparing artifacts on Azure Windows.
`%GITHUB_OUTPUT%` is only defined and used for GitHub Actions, but the
fact that it is not defined causes a syntax error when the script is
run on Azure (since the file to redirect output to is empty). Switching
to delayed expansion with `!GITHUB_OUTPUT!` fixes the syntax error
because those statements are behind an if statement that is not entered
when run on Azure.