Prevent slowdown in yum due to high file descriptor count
Set the maximum number of file descriptors to 1024 before going through the `yum
install` step of the build, only in case `yum_requirements.txt` is present in
the recipe. This is done to mitigate a bug with old versions of rpm such as the
one shipped with the Centos7 container. See https://bugzilla.redhat.com/show_bug.cgi?id=1537564