Allow and validate `schema_version` key in v1 recipes (#2207)
* Accept `schema_version` key in v1 recipes

Allow the v1 recipes to contain a top-level `schema_version` key,
and expect it to be ordered first.  The key is defined in CEP 14:
https://conda.org/learn/ceps/cep-0014#schema-version

Fixes #2201

* Add `schema_version` validation

For the time being, just verify that `schema_version` is 1 when using
`recipe.yaml` format, and refuse to lint if it is not.  In the future,
the code will probably need to be adapted to change `recipe_version`
based on `schema_version`.

* Add a news entry

---------

Co-authored-by: Matthew R. Becker <beckermr@users.noreply.github.com>