Fix handling package version 0, refactor and deduplicate
Fix the boolean logic that caused version of 0 (that is a valid package
version) to be incorrectly interpreted as missing, and emit linter
errors.  While at it, deduplicate the recipe v0 linter to reuse
the utility code added along with v1 lints, and refactor the v1 linter
to reduce the number of redundant conditions and make the logic easier
to follow.

Fixes #2215