Support running `conda-smithy lint` in feedstock directory
Extend the default behavior of `conda-smithy lint` to detect
if a feedstock directory has been passed in place of the recipe
directory (e.g. by running it with no paths specified), and handle
the paths appropriately.

The new logic covers three possible scenarios:

1. If `--feedstock-directory` is passed, everything works as before.

2. If not, the specified directory is checked for `meta.yaml`
   and `recipe.yaml`, also as before.

3. If neither exists, the specified directory is checked for
   `conda-forge.yml`.  If it exists, it set to be the feedstock
   directory, and the file is parsed to determine the correct recipe
   subdirectory.

This is primarily meant to address my common mistake of running:

    conda smithy lint

in the feedstock directory, which can lead to pretty confusing error
messages, particularly if the feedstock is using v1 recipes, and smithy
says it can't find `recipe/meta.yaml` -- and you start wondering whether
you've made a typo in `conda-forge.yml` or what.

This is an alternative to #2249.