GH-hosted runners: automatically resize Windows partitions to max
Automatically resize all partitions on Windows running to use all
the available space.  This is necessary to actually use the additional
space provided on CIrun runners.  The code is copied verbatim from
conda-forge/pytorch-cpu-feedstock where it was needed for a while now.
Resizing is performed unconditionally, since there is no harm in doing
it even if there is no extra space -- it will simply be a no-op then.

Fixes #2394

Signed-off-by: Michał Górny <mgorny@quansight.com>